requirements.txt
requests==2.21.0
ruamel.yaml==0.15.91
semantic-version==2.6.0
pandoc-include==0.3.2
xd-cwl-utils==0.1.2
